requirements.txt
crewai
langchain_groq
pymongo
requests
python-dotenv
beautifulsoup4
duckduckgo-search
python-telegram-bot
